- fixed: ClusterMember: eventClusterMemberRoleInvalid():  * avoid RC by using only synchronized access  * detect and avoid repeated calls - fixed: ClusterMember: removed setLAllinksActivation() - added: ClusterMember: sendClusterBroadcast() supports checks of link activation before a packet is really sent - fixed: CoordinatorAsClusterMember: eventCoordinatorInvalidation()  * avoid RC by using only synchronized access  * detect and avoid repeated calls - added: Elector: major patch set for providing always a deterministic result  * added: some constant values to make the code more readable, improved debug output  * added: eventReset()  * fixed: avoid dead lock in startElection() by avoid synchronization during access  * added: recheckLocalClusterIsAllowedToWin()  * fixed: avoid RC -> no "synchronized" for checkForWinner()  * fixed: hasSourceHigherPrioriorityThan() interprets an undefined priority as very low value instead of return always false  * added: toString() shows if the parent control entity is active or not - fixed: Elector: removed all old references to "Bully" - added: HRMController:  * track the top coordinator placement in order to provide better validation checks  * detect repeated calls to registerCoordinatorProxy()/unregisterCoordinatorProxy()/registerCoordinator/unregisterCoordinator()  * synchronized isGUIFormerCoordiantorID() in order to avoid half-finished unregisterCoordinator()-processes  * avoid endless loop in createComSession() by using HRMControllerProcesser::isValid()  * auto. reset message numbering in eventSimulationRestarted()  * reworked updates of connectivity priority -> use HRMControllerProcessor for asynchronous processing in order to avoid deadlocks in the packet handling  * avoid deadlock in autoExitSimulation() by deactivating all processors first

+++ fog.routing.hrm/src/de/tuilmenau/ics/fog/packets/hierarchical/topology/AnnounceCoordinator.java
@@ -94,7 +94,7 @@ public class AnnounceCoordinator extends SignalingMessageHrm implements ISignali
 	/**
 	 * Stores the current TTL value. If it reaches 0, the packet will be dropped
 	 */
-	private int mTTL = HRMConfig.Hierarchy.RADIUS;
+	private long mTTL = HRMConfig.Hierarchy.RADIUS;
 	
 	/**
 	 * Stores the logical hop count for the stored route 
@@ -352,7 +352,7 @@ public class AnnounceCoordinator extends SignalingMessageHrm implements ISignali
 	 * 
 	 * @return the TTL value
 	 */
-	public int getTTL()
+	public long getTTL()
 	{
 		return mTTL;
 	}
